+++ packages/skeleton-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @skeletonlabs/skeleton-react
 
+## 1.0.0-next.8
+
+### Patch Changes
+
+- chore: Modified default styles based on theme gen improvements ([#2852](https://github.com/skeletonlabs/skeleton/pull/2852))
+
 ## 1.0.0-next.7
 
 ### Minor Changes