+++ packages/skeleton-svelte/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @skeletonlabs/skeleton-svelte
 
+## 1.0.0-next.10
+
+### Patch Changes
+
+- chore: Modified default styles based on theme gen improvements ([#2852](https://github.com/skeletonlabs/skeleton/pull/2852))
+
 ## 1.0.0-next.9
 
 ### Patch Changes